refactor: public accessors for private GameState attrs, inline StatsService into Analytics (#109 #111)
