--- custom_components/quizify/services/__init__.py
@@ -1,1 +1,0 @@
-"""Services module for Quizify."""
